+++ ports/esp32/common-thingz/thingz_display/Raw/thingz_display_raw_text.h
@@ -7,10 +7,10 @@
 
 typedef struct _thingz_display_raw_text_obj_t {
     mp_obj_base_t base;
-    uint8_t x;
-    uint8_t y;
-    uint8_t screen_x;
-    uint8_t screen_y;
+    int16_t x;
+    int16_t y;
+    int16_t screen_x;
+    int16_t screen_y;
     uint8_t screen_height;
     uint8_t screen_width;
     uint8_t show;
